[REFACTOR] 디렉토리 정리 - 선택한 돌봄 항목 항목별 pk 값 리스트 반환
